Sin/cos with floats.

+++ math.h
@@ -1,7 +1,7 @@
 #ifndef INC_08_PADS_MATH_H
 #define INC_08_PADS_MATH_H
 
-extern int sin(int x);
-extern int cos(int x);
+extern float sin(float x);
+extern float cos(float x);
 
 #endif //INC_08_PADS_MATH_H
